move videobridgeprovider
