+++ src/main/java/org/jitsi/videobridge/rest/health/Health.java
@@ -16,7 +16,7 @@
 
 package org.jitsi.videobridge.rest.health;
 
-import org.jitsi.videobridge.rest.*;
+import org.jitsi.videobridge.util.*;
 
 import javax.servlet.http.*;
 import javax.ws.rs.*;
