+++ src/main/java/org/jitsi/videobridge/rest/health/HealthApp.java
@@ -17,8 +17,8 @@
 package org.jitsi.videobridge.rest.health;
 
 import org.glassfish.jersey.server.*;
-import org.jitsi.videobridge.rest.*;
 import org.jitsi.videobridge.rest.debug.*;
+import org.jitsi.videobridge.util.*;
 import org.osgi.framework.*;
 
 /**
